requirements.txt
scipy==1.4.1
numpy==1.19.2
matplotlib==3.2.1
opencv_python==4.2.0.34
tensorflow_gpu==2.5.0
tqdm==4.46.1
Pillow
h5py
protobuf==3.20.0
easygui
pymongo
pandas
boto3